success Ecode sign down

+++ lot_Ecode/StringtoHex.c
@@ -1,6 +1,6 @@
 #include "StringtoHex.h"
 
-void StringtoTen(char* buff,long int* hex, int n)
+void StringtoTen(char* buff,int* hex, int n)
 {
 	int i;
 	char a[2] = {'0', '\0'};
